requirements.txt
keboola.component
keboola.utils
keboola.http-client
mock
freezegun
requests
openai
pydub
